// src/api/generated/model/rewardPhotoDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

export interface RewardPhotoDto {
  /** Unique identifier */
  id: string;
  /** Timestamp of when the record was created */
  createdAt: string;
  /** Timestamp of the last update of the record */
  updatedAt: string;
  /** Presigned URL изображения */
  url: string;
  /** MIME-тип изображения */
  contentType: string;
  /** Размер изображения в байтах */
  sizeBytes: number;
  /** Порядок фото в галерее */
  sortOrder: number;
}